sites/ap/requirements/base.txt
boto3==1.34.109
django-configurations==2.5.1
# Superior versions of django-storages are not compatible with
# ManifestStaticFilesStorage
django-storages==1.14.3
dockerflow==2024.4.2
factory-boy==3.3.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
git+https://github.com/Tiago-Salles/richie@Tiago-Salles/issues/274-upstream-cookiecutter#egg=richie
#richie==2.32.0
sentry-sdk==2.2.1
mysqlclient==2.1.1
# Remove django-cms dep when upgrading to richie 2.25.0
# django-cms>=3.11.0,<4.0.0
# Plugins
djangocms-snippet==3.1.1
git+https://github.com/openfun/django-cms@fun-3.11.6#egg=django-cms